+++ backend/store/etcd/filter_store.go
@@ -4,10 +4,8 @@ import (
 	"context"
 	"errors"
 
-	"github.com/coreos/etcd/clientv3"
-	"github.com/gogo/protobuf/proto"
+	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/store"
-	"github.com/sensu/sensu-go/types"
 )
 
 var (
@@ -15,7 +13,7 @@ var (
 	eventFilterKeyBuilder  = store.NewKeyBuilder(eventFiltersPathPrefix)
 )
 
-func getEventFilterPath(filter *types.EventFilter) string {
+func getEventFilterPath(filter *corev2.EventFilter) string {
 	return eventFilterKeyBuilder.WithResource(filter).Build(filter.Name)
 }
 
@@ -30,68 +28,44 @@ func (s *Store) DeleteEventFilterByName(ctx context.Context, name string) error 
 		return &store.ErrNotValid{Err: errors.New("must specify name of filter")}
 	}
 
-	resp, err := s.client.Delete(ctx, GetEventFiltersPath(ctx, name))
+	err := Delete(ctx, s.client, GetEventFiltersPath(ctx, name))
 	if err != nil {
-		return &store.ErrInternal{Message: err.Error()}
+		if _, ok := err.(*store.ErrNotFound); ok {
+			err = nil
+		}
 	}
-
-	if resp.Deleted != 1 {
-		return &store.ErrNotFound{Key: name}
-	}
-
-	return nil
+	return err
 }
 
 // GetEventFilters gets the list of filters for a namespace.
-func (s *Store) GetEventFilters(ctx context.Context, pred *store.SelectionPredicate) ([]*types.EventFilter, error) {
-	filters := []*types.EventFilter{}
+func (s *Store) GetEventFilters(ctx context.Context, pred *store.SelectionPredicate) ([]*corev2.EventFilter, error) {
+	filters := []*corev2.EventFilter{}
 	err := List(ctx, s.client, GetEventFiltersPath, &filters, pred)
 	return filters, err
 }
 
 // GetEventFilterByName gets an EventFilter by name.
-func (s *Store) GetEventFilterByName(ctx context.Context, name string) (*types.EventFilter, error) {
+func (s *Store) GetEventFilterByName(ctx context.Context, name string) (*corev2.EventFilter, error) {
 	if name == "" {
 		return nil, &store.ErrNotValid{Err: errors.New("must specify name of filter")}
 	}
 
-	resp, err := s.client.Get(ctx, GetEventFiltersPath(ctx, name))
-	if err != nil {
-		return nil, &store.ErrInternal{Message: err.Error()}
-	}
-	if len(resp.Kvs) == 0 {
-		return nil, nil
-	}
-
-	filterBytes := resp.Kvs[0].Value
-	filter := &types.EventFilter{}
-	if err := unmarshal(filterBytes, filter); err != nil {
-		return nil, &store.ErrDecode{Err: err}
+	var filter corev2.EventFilter
+	if err := Get(ctx, s.client, GetEventFiltersPath(ctx, name), &filter); err != nil {
+		if _, ok := err.(*store.ErrNotFound); ok {
+			err = nil
+		}
+		return nil, err
 	}
 
-	return filter, nil
+	return &filter, nil
 }
 
 // UpdateEventFilter updates an EventFilter.
-func (s *Store) UpdateEventFilter(ctx context.Context, filter *types.EventFilter) error {
+func (s *Store) UpdateEventFilter(ctx context.Context, filter *corev2.EventFilter) error {
 	if err := filter.Validate(); err != nil {
 		return &store.ErrNotValid{Err: err}
 	}
 
-	filterBytes, err := proto.Marshal(filter)
-	if err != nil {
-		return &store.ErrEncode{Err: err}
-	}
-
-	cmp := clientv3.Compare(clientv3.Version(getNamespacePath(filter.Namespace)), ">", 0)
-	req := clientv3.OpPut(getEventFilterPath(filter), string(filterBytes))
-	res, err := s.client.Txn(ctx).If(cmp).Then(req).Commit()
-	if err != nil {
-		return &store.ErrInternal{Message: err.Error()}
-	}
-	if !res.Succeeded {
-		return &store.ErrNamespaceMissing{Namespace: filter.Namespace}
-	}
-
-	return nil
+	return CreateOrUpdate(ctx, s.client, getEventFilterPath(filter), filter.Namespace, filter)
 }